Change XAML windows to pages and creates a main page with initial formatting

+++ NumberConversion/NumberConvert.cs
@@ -10,17 +10,17 @@ namespace NumberConversion
             if (toBase == 16)
             {
                 //TODO Converting from decimal to hex doesn't seem to force the correct number of digits
-                return Convert.ToInt32(value, fromBase).ToString($"X{length}");
+                return Convert.ToInt64(value, fromBase).ToString($"X{length}");
             }
 
             else if(toBase == 10)
             {
-                return Convert.ToInt32(value, fromBase).ToString("D");
+                return Convert.ToInt64(value, fromBase).ToString("D");
             }
 
             else if (toBase == 2)
             {
-                String bitVal = Convert.ToString(Convert.ToInt32(value, fromBase), 2);
+                String bitVal = Convert.ToString(Convert.ToInt64(value, fromBase), 2);
                 int num0 = length - bitVal.Length; // the number of zeros that need to be prepended to reach desired length
                 num0 = Math.Max(0, num0); // dont want to append negative values
                 string prepend = new string('0', num0);
@@ -38,7 +38,7 @@ namespace NumberConversion
             {
                 //twos compliment can be written as maxvalue - value
                 int maxValue = (int) Math.Pow(2, value.Length);
-                int twoCompDecimalNum = maxValue - Convert.ToInt32(value,2);
+                long twoCompDecimalNum = maxValue - Convert.ToInt64(value,2);
                 String bitVal= ConvertToBase(twoCompDecimalNum.ToString(), 10, 2);
                 int num0 = length - bitVal.Length; // the number of zeros that need to be prepended to reach desired length
                 num0 = Math.Max(0, num0); // dont want to append negative values
